option_pricing_project/requirements.txt
asgiref==3.8.1
contourpy==1.1.0
cycler==0.11.0
Django==4.2.11
fonttools==4.40.0
importlib_resources==6.4.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
zipp==3.18.1
